RDRP-859: moved ultfoc mapping to own script
